feat(revpi64-bookworm): Use the bookworm DEB 822 sources
The 'revpi-repo' is available for bookworm in DEB 822 format. The links
to the files have been adapted and the revpi.gpg is stored in
/usr/share/keyrings. However, by calling ADD, this file is only
readable by the root user, which leads to an error when using (apt).
The files that contain the public keys must also be readable by group
and others (default for the keyrings folder).

Signed-off-by: Sven Sager <akira@narux.de>