feat: Copy artifacts, if --git-export-dir argument is set
In rare cases, it can happen that the build on the mapped file system of
the host system fails. This could happen, for example, through changed
permissions. To circumvent this, e.g. for gbp, the parameter
git-export-dir can be specified. The packages are then built in the
file system by the Docker container. The resulting packets must be
copied from the container file system to the host file system at the
end.

Signed-off-by: Sven Sager <akira@narux.de>